requirements.txt
pyjwt
pyjwt[crypto]